fix: 경매 목록 grid 오류 수정
